240111 1차 수정
1. 경섭님이 작업해주신 회피기능 체리픽

2. Program.cs에서 플레이어 / 아이템 / 몬스터 클래스 따로 분리해서 정리

3. 플레이어가 이미 죽은 몬스터를 공격할 경우 다시 입력을 받게 수정
