+++ Program.cs
@@ -4,368 +4,6 @@ using System.Threading;
 
 namespace _5week_assignment
 {
-    public class Character
-    {
-        public enum ClassType
-        {
-            None = 0,
-            Knight = 1,     //전사
-            Archer = 2,     // 궁수
-            Mage = 3        //마법사
-        }
-        public string Name { get; set; }
-        public string Job{ get; set; }
-        public int Level { get; set; }
-        public int Atk { get; set; }
-        public int Def { get; set; }
-        public int Hp { get; set; }
-
-        public bool isDead { get; set; }
-        public int currentHP { get; set; }
-        public int Gold { get; set; }
-
-        public void CreatePlayer()     // 플레이어 생성
-        {
-            Console.WriteLine("이름을 입력해주세요.");
-            Console.Write(">>");
-            string input = Console.ReadLine();
-
-            Name = input;
-            Gold = 1500;
-            isDead = false;
-            Level = 1;
-            Console.Clear();
-            ChoiceClass();
-            Console.Clear();
-        }
-
-        public ClassType ChoiceClass()  // 직업선택 메서드
-        {
-            Console.WriteLine("직업을 선택하세요!");
-            Console.WriteLine("[1] 전사");
-            Console.WriteLine("[2] 궁수");
-            Console.WriteLine("[3] 마법사");
-            ClassType choice = ClassType.None;
-
-            switch(CheckValidInput(1,3))
-            {
-                case 1:
-                    choice = ClassType.Knight;
-                    Job = "전사";
-                    Hp = 100;
-                    currentHP = Hp;
-                    Atk = 10;
-                    Def = 20;
-                    break;
-                case 2:
-                    choice = ClassType.Archer;
-                    Job = "궁수";
-                    Hp = 80;
-                    currentHP = Hp;
-                    Atk = 15;
-                    Def = 15;
-                    break;
-                case 3:
-                    choice = ClassType.Mage;
-                    Job = "마법사";
-                    Hp = 60;
-                    currentHP = Hp;
-                    Atk = 20;
-                    Def = 8;
-                    break;
-                default:
-                    Console.Clear();
-                    ChoiceClass();
-                    break;
-
-            }
-            return choice;    
-        }
-
-        //playerattackdamage
-        public void PlayerAttack(Monster monster, out int damaged)
-        {
-            Random rand = new Random();
-            int minAtk = Atk - (int)Math.Ceiling(Atk * 0.1);
-            int maxAtk = Atk + (int)Math.Ceiling(Atk * 0.1);
-            int attack = rand.Next(minAtk, maxAtk + 1);
-            //10%확률
-            int hit = rand.Next(1, 101);
-            if (hit <= 90)
-            { 
-                attack = 0;
-                damaged = attack;
-            }
-            else
-            {
-                monster.currentHp -= attack;
-                damaged = attack;
-                //치명타코드
-                
-            }
-
-        }
-
-        private int CheckValidInput(int min, int max)
-        {
-            int keyInput;
-            bool result;
-
-            do
-            {
-                Console.WriteLine("직업 번호를 입력하세요!");
-                Console.Write(">>");
-                result = int.TryParse(Console.ReadLine(), out keyInput);
-            } while (result == false || CheckIfValid(keyInput, min, max) == false);
-
-            return keyInput;
-
-        }
-
-        private bool CheckIfValid(int keyInput, int min, int max)
-        {
-            if (min <= keyInput && keyInput <= max)
-            {
-                return true;
-            }
-            else
-            {
-                return false;
-            }
-        }
-
-        public void PlayerInfo()
-        {
-
-            Console.WriteLine($"Lv.{Level.ToString("00")} {Name} ({Job})");
-            Console.WriteLine($"HP {Hp}/{currentHP}");
-        }
-
-        #region 회피코드였던것
-        //int num = rand.Next(0, 101);
-        //float[] p = { 10f, 15f, 75f };
-
-        //float cumulative = 0f;
-        //int target = -1;
-        //for (int i = 0; i < 3; i++)
-        //{
-        //    cumulative += p[i];
-        //    if (num <= cumulative)
-        //    {
-        //        target = 2 - i;
-        //        break;
-        //    }
-        //}
-        //switch(target)
-        //{
-        //    case 0:
-        //        int nodamage = 0;
-        //        damaged = nodamage;
-        //        break;
-        //    case 1:
-        //        //치명타
-        //    case 2:
-        //        damaged = attack;
-        //        break;
-        //}
-        #endregion
-    }
-
-    
-    #region 아이템
-    //public class Item
-    //{
-    //    public string Name { get; }
-    //    public string Description { get; }
-    //    public int Type { get; }
-    //    public int Atk { get; }
-    //    public int Def { get; }
-    //    public int Hp { get; }
-    //    public bool isEquipped { get; set; }
-
-    //    public Item(string name, string description, int type, int atk, int def, int hp, bool isEquipped = false)
-    //    {
-    //        Name = name;
-    //        Description = description;
-    //        Type = type;
-    //        Atk = atk;
-    //        Def = def;
-    //        Hp = hp;
-    //        isEquipped = isEquipped;
-    //    }
-
-    //    public static int ItemCnt = 0;
-
-    //    public void PrintItemStatDescription(bool withNumber = false, int idx = 0)
-    //    {
-    //        Console.Write("- ");
-
-    //        if (withNumber)
-    //        {
-    //            Console.ForegroundColor = ConsoleColor.DarkMagenta;
-    //            Console.Write("{0} ", idx);
-    //            Console.ResetColor();
-    //        }
-
-    //        if (isEquipped)
-    //        {
-    //            Console.Write("[");
-    //            Console.ForegroundColor = ConsoleColor.Cyan;
-    //            Console.Write("E");
-    //            Console.ResetColor();
-    //            Console.Write("]");
-    //        }
-    //        else
-    //        {
-    //            Console.Write(PadRightForMixedText(Name, 12));
-    //        }
-
-    //        Console.Write(" | ");
-
-    //        if (Atk != 0)
-    //        {
-    //            Console.Write($"Atk {(Atk >= 0 ? "+" : "")}{Atk}");
-    //        }
-    //        if (Def != 0)
-    //        {
-    //            Console.Write($"Def {(Def >= 0 ? "+" : "")}{Def}");
-    //        }
-    //        if (Hp != 0)
-    //        {
-    //            Console.Write($"Hp {(Hp >= 0 ? "+" : "")}{Hp}");
-    //        }
-
-    //        Console.Write(" | ");
-
-    //        Console.WriteLine(Description);
-
-    //    }
-
-    //    public static int GetPrintableLength(string str)
-    //    {
-    //        int length = 0;
-    //        foreach (char c in str)
-    //        {
-    //            if (char.GetUnicodeCategory(c) == System.Globalization.UnicodeCategory.OtherLetter)
-    //            {
-    //                length += 2;
-    //            }
-    //            else
-    //            {
-    //                length += 1;
-    //            }
-    //        }
-
-    //        return length;
-    //    }
-
-    //    public static string PadRightForMixedText(string str, int totalLength)
-    //    {
-    //        int currentLength = GetPrintableLength(str);
-    //        int padding = totalLength - currentLength;
-    //        return str.PadRight(str.Length + padding);
-    //    }
-    //}
-    #endregion
-
-    public class Monster
-    {
-        enum MonsterType
-        {
-            None,
-            LeeHanSol,
-            MonYeongOh,
-            HanHyoseung,
-        }
-
-        public string Name;
-        public int Hp;
-        public int currentHp;
-        public int Atk;
-        public int Def;
-        public int Level;
-
-        public bool isDead; // 죽었니 살았니?
-
-        public Monster()
-        {
-            Random rand = new Random();
-            int randMonster = rand.Next(1,4);
-            switch(randMonster)
-            {
-                case (int)MonsterType.LeeHanSol:
-                    Name = "이한솔 매니저";
-                    Level = 2;
-                    Hp = 10;
-                    currentHp = Hp;
-                    Atk = 3;
-                    Def = 0;
-                    break;
-                case (int)MonsterType.MonYeongOh:
-                    Name = "문영오 매니저";
-                    Level = 3;
-                    Hp = 15;
-                    currentHp = Hp;
-                    Atk = 6;
-                    Def = 1;
-                    break;
-                case (int)MonsterType.HanHyoseung:
-                    Name = "한효승 매니저";
-                    Level = 5;
-                    Hp = 25;
-                    currentHp = Hp;
-                    Atk = 9;
-                    Def = 3;
-                    break;
-            }
-
-        }
-
-        public void MonsterInfo(bool withNumber = false, int index = 0)
-        {
-            Console.Write("- ");
-            
-            if(isDead)
-            {
-                Console.ForegroundColor = ConsoleColor.DarkGray;
-                Console.WriteLine($"Lv.{Level.ToString("00")} {Name}  Dead");
-                Console.ResetColor();
-            }
-            else if (withNumber)
-            {
-                Console.Write($"{index}  ");
-                Console.WriteLine($"Lv.{Level.ToString("00")} {Name} HP {currentHp}");
-            }
-            else
-            {
-                Console.WriteLine($"Lv.{Level.ToString("00")} {Name} HP {currentHp}");
-            }
-            
-        }
-
-        public void MonsterAttack(Character character,out int damaged)
-        {
-            Random rand = new Random();
-            int minAtk = Atk - (int)Math.Ceiling(Atk * 0.1);
-            int maxAtk = Atk + (int)Math.Ceiling(Atk * 0.1);
-            int attack = rand.Next(minAtk, maxAtk + 1);
-
-            //10%확률
-            int hit = rand.Next(1, 101);
-            if (hit <= 90)
-            {
-                attack = 0;
-                damaged = attack;
-            }
-            else
-            {
-                character.currentHP -= attack;
-                damaged = attack;
-                //치명타코드
-            }
-        }
-    }
-
     internal class Program
     {
         private static Character _player = new Character();
@@ -511,7 +149,12 @@ namespace _5week_assignment
             Console.WriteLine();
             Console.WriteLine("0. 취소");
 
+
             int input = CheckValidInput(0, monsterPool.Count);
+            while (monsterPool[input-1].isDead)
+            {
+                input = CheckValidInput(0,monsterPool.Count);
+            }
             switch (input)
             {
                 case 0:
